Feat: removed tooltip, added modal component
- removed tooltip. shadcn sucks. It is a error generator
- shadcn sucks. Therefore, I made my own modal.
