+++ pages/404.tsx
@@ -1,6 +1,6 @@
 export default function Custom404() {
   return (
-    <div className='w-full flex flex-col items-center justify-center'>
+    <div className='w-full h-full flex flex-col items-center justify-center'>
       <h1 className='text-5xl font-bold'>404 - 🤪</h1>
     </div>
   );
